+++ src/layer/domain/indexeddb/model/channel.ts
@@ -65,6 +65,11 @@ export class ChannelStore<K extends string, V extends StoreChannelObject<K>> {
     }
   }
   private readonly schema: Schema<K, V>;
+  public readonly events_ = {
+    load: new Observable<never[] | [K] | [K, keyof V | ''] | [K, keyof V | '', ChannelStore.EventType], ChannelStore.Event<K, V>, void>(),
+    save: new Observable<never[] | [K] | [K, keyof V | ''] | [K, keyof V | '', ChannelStore.EventType], ChannelStore.Event<K, V>, void>(),
+    loss: new Observable<never[] | [K] | [K, keyof V | ''] | [K, keyof V | '', ChannelStore.EventType], ChannelStore.Event<K, V>, void>()
+  };
   public readonly events = {
     load: new Observable<never[] | [K] | [K, keyof V | ''] | [K, keyof V | '', ChannelStore.EventType], ChannelStore.Event<K, V>, void>(),
     save: new Observable<never[] | [K] | [K, keyof V | ''] | [K, keyof V | '', ChannelStore.EventType], ChannelStore.Event<K, V>, void>(),
@@ -146,8 +151,11 @@ class Schema<K extends string, V extends StoreChannelObject<K>> {
     const keys = this.data ? this.data.keys() : [];
 
     this.data = new DataStore<K, V>(this.store_.name, this.attrs_);
+    void this.data.events.load.monitor([], ev => this.store_.events_.load.emit([ev.key, ev.attr, ev.type], ev));
     void this.data.events.load.monitor([], ev => this.store_.events.load.emit([ev.key, ev.attr, ev.type], ev));
+    void this.data.events.save.monitor([], ev => this.store_.events_.save.emit([ev.key, ev.attr, ev.type], ev));
     void this.data.events.save.monitor([], ev => this.store_.events.save.emit([ev.key, ev.attr, ev.type], ev));
+    void this.data.events.loss.monitor([], ev => this.store_.events_.loss.emit([ev.key, ev.attr, ev.type], ev));
     void this.data.events.loss.monitor([], ev => this.store_.events.loss.emit([ev.key, ev.attr, ev.type], ev));
     this.access = new AccessStore<K>(this.store_.name, this.data.events_.access);
     this.expire = new ExpiryStore<K>(this.store_.name, this.store_, this.data.events_.access, this.expiries_, this.cancelable_);